backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-community
langchain_huggingface
sentence_transformers
openai
faiss-cpu
tiktoken
python-dotenv